redesign manifest custom availability check test
